chore: remote cli

+++ core/container/run.ts
@@ -1,9 +1,10 @@
 import { z } from "zod";
 import kubernetes from "util/kube-client.ts";
-import { Context } from "core/context.ts";
+import { Context, StdioContext } from "core/context.ts";
 import { Pod } from "@cloudydeno/kubernetes-apis/core/v1";
 import { WatchEvent } from "@cloudydeno/kubernetes-apis/common.ts";
-import attach from './attach.ts'
+import attach from "./attach.ts";
+import { stdout } from "node:process";
 
 export const meta = {
   aliases: {
@@ -14,28 +15,26 @@ export const meta = {
   },
 };
 
-export const Input = z.tuple([
-  z.string().describe("Name of the container"),
-  z.object({
-    image: z.string().describe("Container image to run"),
-    env: z.array(z.string()).optional().describe("Set environment variables"),
-    port: z.array(z.number()).optional().describe("port to expose"),
-    interactive: z.boolean().optional().default(false).describe("Run interactively"),
-    terminal: z.boolean().optional().default(false).describe("Run in a terminal"),
-    // scaleType: z.enum(["concurrency", "rps", "cpu", "memory"]).optional().describe("Type of scaling to apply"),
-    // scale: z.object({
-    //   min: z.number().optional(),
-    //   max: z.number().optional(),
-    //   target: z.number().optional(),
-    // }).optional().describe("Scaling configuration"),
-    command: z.string().optional().describe("Command to run in the container"),
-  }),
-]);
+export const Input = z.object({
+  name: z.string().describe("Name of the container"),
+  image: z.string().describe("Container image to run"),
+  env: z.array(z.string()).optional().describe("Set environment variables"),
+  port: z.array(z.number()).optional().describe("port to expose"),
+  interactive: z.boolean().optional().default(false).describe("Run interactively"),
+  terminal: z.boolean().optional().default(false).describe("Run in a terminal"),
+  // scaleType: z.enum(["concurrency", "rps", "cpu", "memory"]).optional().describe("Type of scaling to apply"),
+  // scale: z.object({
+  //   min: z.number().optional(),
+  //   max: z.number().optional(),
+  //   target: z.number().optional(),
+  // }).optional().describe("Scaling configuration"),
+  command: z.string().optional().describe("Command to run in the container"),
+});
 
 export type Input = z.infer<typeof Input>;
 
-export default (context: Context) => async (input: Input) => {
-  const [name, { image, env = [], port = [], interactive, terminal, command }] = input;
+export default (context: StdioContext) => async (input: Input) => {
+  const { name, image, env = [], port = [], interactive, terminal, command } = input;
 
   const podResource: Pod = {
     apiVersion: "v1",
@@ -85,54 +84,52 @@ export default (context: Context) => async (input: Input) => {
     },
   };
 
-  // Delete existing pod if it exists
-  {
-    const pod = await kubernetes.CoreV1.namespace("default").getPod(name).catch(() => null);
-    if (pod) {
-      console.warn(`Container ${name} already exists. Deleting it before creating a new one.`);
-      kubernetes.CoreV1.namespace("default").deletePod(name, {
+  const stdoutWriter = context.stdio.stdout.getWriter();
+  const stderrWriter = context.stdio.stderr.getWriter();
+
+  // Check if the pod already exists and is running
+  const pod = await kubernetes.CoreV1.namespace("default").getPod(name).catch(() => null);
+  if (!pod || pod.status?.phase !== "Running") {
+    // Delete existing pod if not running
+    {
+      if (pod && pod.status?.phase !== "Running") {
+        stderrWriter.write(`Container ${name} already exists. Deleting it before creating a new one.\r\n`);
+        kubernetes.CoreV1.namespace("default").deletePod(name, {
+          abortSignal: context.signal,
+        });
+        stdoutWriter.write(`Waiting for container ${name} to be deleted...\r\n`);
+        // Wait for the pod to be deleted
+        await waitPodEvent(context, name, "DELETED");
+      }
+    }
+
+    // Create pod if it doesn't exist or is not running
+    {
+      kubernetes.CoreV1.namespace("default").createPod(podResource, {
         abortSignal: context.signal,
       });
-      console.info(`Waiting for container ${name} to be deleted...`);
-      // Wait for the pod to be deleted
-      await waitPodEvent(context, name, "DELETED");
+      stdoutWriter.write(
+        `Container ${name} created. Waiting for it to be ready...\r\n`,
+      );
+      // Watch for pod events
+      await waitPodStatus(context, name, "ContainersReady").catch(() => {});
     }
+  } else {
+    stderrWriter.write(
+      `Container ${name} is already running.\r\n`,
+    );
+    return;
   }
 
-  // Create pod
-  {
-    kubernetes.CoreV1.namespace("default").createPod(podResource, {
-      abortSignal: context.signal,
-    });
-    console.info(`Container ${name} created. Waiting for it to be ready...`);
-    // Watch for pod events
-    await waitPodStatus(context, name, "ContainersReady");
-  }
-
-  // Delete the pod if the context is aborted
-  context.signal?.addEventListener("abort", async () => {
-    // Delete the pod if the context is aborted
-    kubernetes.CoreV1.namespace("default").deletePod(name, {
-      abortSignal: context.signal,
-    });
-    // Wait for the pod to be deleted
-    await waitPodEvent(context, name, "DELETED");
-  });
+  stdoutWriter.releaseLock();
+  stderrWriter.releaseLock();
 
   // Attach to the pod if interactive or terminal mode is enabled
-  await attach(context)([name, {
+  await attach(context)({
+    name,
     interactive,
     terminal,
-  }])
-
-  // Delete the pod
-  {
-    kubernetes.CoreV1.namespace("default").deletePod(name, {
-      abortSignal: context.signal,
-    });
-    // Wait for the pod to be deleted
-    await waitPodEvent(context, name, "DELETED");
-  }
+  });
 };
 
 async function waitPodEvent(context: Context, name: string, eventType: WatchEvent<any, any>["type"]): Promise<void> {
